Day One
This is my first day's progression learning HTML. Today wasn't too difficult and I passed a lot of the tests that i had to pass for writing code on my on. Right now i am creating a Cat Photo App, but i'm going to take a break and start learning Java script.. Until the next entry.